Added service interfaces to backend
