--- src/main/java/com/globallogic/amcr/service/Service.java
@@ -1,4 +1,0 @@
-package com.globallogic.amcr.service;
-
-public interface Service {
-}
